Changed notification regex to work on Gentoo
Typically X is started on Gentoo as /usr/bin/X which is a symlink to
Xorg binary. It shows up on pgrep output as just "/usr/bin/X", so
looking for Xorg does not find the running process.

This change allows pgrep to locate X if it's just running as "X",
making "org" and "wayland" matches optional, which in turn, fixes
notifications.